+++ dpctl/_sycl_device.pyx
@@ -27,6 +27,9 @@ from ._backend cimport (  # noqa: E211
     DPCTLDefaultSelector_Create,
     DPCTLDevice_AreEq,
     DPCTLDevice_CanAccessPeer,
+    DPCTLDevice_CanCompileOpenCL,
+    DPCTLDevice_CanCompileSPIRV,
+    DPCTLDevice_CanCompileSYCL,
     DPCTLDevice_Copy,
     DPCTLDevice_CreateFromSelector,
     DPCTLDevice_CreateSubDevicesByAffinity,
@@ -99,7 +102,6 @@ from ._backend cimport (  # noqa: E211
     DPCTLDevice_GetPreferredVectorWidthShort,
     DPCTLDevice_GetProfilingTimerResolution,
     DPCTLDevice_GetSingleFPConfig,
-    DPCTLDevice_GetSubGroupIndependentForwardProgress,
     DPCTLDevice_GetSubGroupSizes,
     DPCTLDevice_GetVendor,
     DPCTLDevice_GetVendorId,
@@ -1245,20 +1247,6 @@ cdef class SyclDevice(_SyclDevice):
             DPCTLDevice_GetMaxNumSubGroups(self._device_ref)
         )
         return max_num_sub_groups
-
-    @property
-    def sub_group_independent_forward_progress(self):
-        """ Returns ``True`` if the device supports independent forward progress
-        of sub-groups with respect to other sub-groups in the same work-group.
-
-        Returns:
-            bool:
-                Indicates if the device supports independent forward progress
-                of sub-groups.
-        """
-        return DPCTLDevice_GetSubGroupIndependentForwardProgress(
-            self._device_ref
-        )
 
     @property
     def sub_group_sizes(self):
@@ -2868,6 +2856,34 @@ cdef class SyclDevice(_SyclDevice):
         if dev_id < 0:
             raise ValueError("device could not be found")
         return dev_id
+
+    cpdef bint can_compile(self, str language):
+        """
+        Check whether it is possible to create an executable kernel_bundle
+        for this device from the given source language.
+
+        Parameters:
+            language
+                Input language. Possible values are "spirv" or "spv" for
+                SPIR-V binary files, "opencl" or "ocl" for OpenCL C device code
+                and "sycl" for SYCL device code.
+
+        Returns:
+            bool:
+                True if compilation is supported, False otherwise.
+
+        Raises:
+            ValueError:
+                If an unknown source language is used.
+        """
+        if language == "spirv" or language == "spv":
+            return DPCTLDevice_CanCompileSPIRV(self._device_ref)
+        if language == "opencl" or language == "ocl":
+            return DPCTLDevice_CanCompileOpenCL(self._device_ref)
+        if language == "sycl":
+            return DPCTLDevice_CanCompileSYCL(self._device_ref)
+
+        raise ValueError(f"Unknown source language {language}")
 
 
 cdef api DPCTLSyclDeviceRef SyclDevice_GetDeviceRef(SyclDevice dev):