+++ libsyclinterface/include/syclinterface/dpctl_sycl_device_interface.h
@@ -458,18 +458,6 @@ DPCTL_API
 __dpctl_give DPCTLDeviceVectorRef DPCTLDevice_CreateSubDevicesByAffinity(
     __dpctl_keep const DPCTLSyclDeviceRef DRef,
     DPCTLPartitionAffinityDomainType PartAffDomTy);
-/*!
- * @brief Wrapper over
- * device.get_info<info::device::sub_group_independent_forward_progress>.
- *
- * @param    DRef           Opaque pointer to a ``sycl::device``
- * @return   Returns true if the device supports independent forward progress of
- * sub-groups with respect to other sub-groups in the same work-group.
- * @ingroup DeviceInterface
- */
-DPCTL_API
-bool DPCTLDevice_GetSubGroupIndependentForwardProgress(
-    __dpctl_keep const DPCTLSyclDeviceRef DRef);
 
 /*!
  * @brief Wrapper over
@@ -1093,5 +1081,41 @@ DPCTL_API
 __dpctl_give int *DPCTLDevice_GetPartitionAffinityDomains(
     __dpctl_keep const DPCTLSyclDeviceRef DRef,
     size_t *res_len);
+
+/*!
+ * @brief Checks whether it is possible to create executables kernel bundles
+ * from SPIR-V binaries on this device.
+ *
+ * @param DRef  Opaque pointer to a ``sycl::device``.
+ * @return True if creation is supported.
+ * #DPCTLSyclDeviceRef objects
+ * @ingroup DeviceInterface
+ */
+DPCTL_API
+bool DPCTLDevice_CanCompileSPIRV(__dpctl_keep const DPCTLSyclDeviceRef DRef);
+
+/*!
+ * @brief Checks whether it is possible to create executables kernel bundles
+ * from OpenCL source code on this device.
+ *
+ * @param DRef  Opaque pointer to a ``sycl::device``.
+ * @return True if creation is supported.
+ * #DPCTLSyclDeviceRef objects
+ * @ingroup DeviceInterface
+ */
+DPCTL_API
+bool DPCTLDevice_CanCompileOpenCL(__dpctl_keep const DPCTLSyclDeviceRef DRef);
+
+/*!
+ * @brief Checks whether it is possible to create executables kernel bundles
+ * from SYCL source code on this device.
+ *
+ * @param DRef  Opaque pointer to a ``sycl::device``.
+ * @return True if creation is supported.
+ * #DPCTLSyclDeviceRef objects
+ * @ingroup DeviceInterface
+ */
+DPCTL_API
+bool DPCTLDevice_CanCompileSYCL(__dpctl_keep const DPCTLSyclDeviceRef DRef);
 
 DPCTL_C_EXTERN_C_END